+++ src/store/src/Bridge/Elasticsearch/Store.php
@@ -68,6 +68,13 @@ final class Store implements ManagedStoreInterface, StoreInterface
         $this->request('DELETE', $this->indexName);
     }
 
+    public function count(): int
+    {
+        $result = $this->request('GET', \sprintf('%s/_count', $this->indexName));
+
+        return $result['count'] ?? 0;
+    }
+
     public function add(VectorDocument|array $documents): void
     {
         if ($documents instanceof VectorDocument) {