+++ src/store/src/Bridge/MongoDb/Store.php
@@ -112,6 +112,11 @@ final class Store implements ManagedStoreInterface, StoreInterface
         $this->getCollection()->drop();
     }
 
+    public function count(): int
+    {
+        return $this->getCollection()->countDocuments();
+    }
+
     public function add(VectorDocument|array $documents): void
     {
         if ($documents instanceof VectorDocument) {