+++ src/store/src/InMemory/Store.php
@@ -107,6 +107,11 @@ class Store implements ManagedStoreInterface, StoreInterface, ResetInterface
         $this->drop();
     }
 
+    public function count(): int
+    {
+        return \count($this->documents);
+    }
+
     /**
      * @param array{maxItems?: positive-int, filter?: callable(VectorDocument): bool} $options
      *