+++ src/store/src/Bridge/MongoDb/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 0.7
 ---
 
+ * Add `count` method
  * Add support for `Store::remove()` method
 
 0.1